+++ supreme/verifier/src/jvmTest/kotlin/examples/ConfigGen.kt
@@ -35,7 +35,7 @@ val ConfigurationExampleGenerator by testSuite(compartment = { TestCompartment.S
         /*(1)!*/
         AndroidAttestationConfiguration.AppData(
             packageName = "at.asitplus.attestation_client",
-            signerFingerprints = listOf("34 b9 76 2c 4d 6c 90 d4 84 31 94 0c 57 bd e7 31 42 58 b2 64 20 ec".parseHex())
+            signerFingerprints = setOf("34 b9 76 2c 4d 6c 90 d4 84 31 94 0c 57 bd e7 31 42 58 b2 64 20 ec".parseHex())
         ),
         verifiedBootKeys = linkedSetOf(
             VerifiedBootKey.OEM,