+++ CHANGELOG.md
@@ -5,9 +5,14 @@ dedicated artefacts,
 this changelog also includes the original WARDEN changelog.
 
 # NEXT
-* Support custom verified boot keys
-    * Allows treating self-signed just like a locked bootloader if desired
-    * Makes it possible to disallow even OEM ROMS and allow only GrapheneOS, for example 
+* Features
+    * Support custom verified boot keys
+        * Allows treating self-signed just like a locked bootloader if desired
+        * Makes it possible to disallow even OEM ROMS and allow only GrapheneOS, for example
+* API Changes:
+    * Signer fingerprints is now a `Set` and not a list anymore
+* Behavioural Changes:
+    * `parseHex` now also strips `:` by default
 
 # 1.0.0-RC5
 * Features